% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animations}
\alias{animations}
\title{List of all available animations}
\format{
An object of class \code{list} of length 7.
}
\usage{
animations
}
\description{
A list containing all available animations.
}
\details{
Source: \url{https://github.com/IanLunn/Hover}
}
\keyword{datasets}
