% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_hover.R
\name{use_hover}
\alias{use_hover}
\title{Use the hover package}
\usage{
use_hover()
}
\description{
Enables hover by including the CSS file necessary for the
animations.
}
\examples{
if (interactive()) {
  library(shiny)
  library(hover)

  ui <- fluidPage(
    use_hover(),
    hover_button(
      inputId = "btn",
      label = "hello hover!",
      icon = icon("refresh"),
      button_animation = "rotate",
      icon_animation = "spin"
    )
  )

  server <- function(input, output, session) {

  }

  shinyApp(ui, server)
}
}
